Abstracted login into Provider class, improved drush validation as per Issue #1.

+++ library/fluxsauce/switchboard/src/Sqlite.php
@@ -7,7 +7,9 @@ namespace Fluxsauce\Switchboard;
 
 class Sqlite {
   public static function siteDelete() {
-    unlink(drush_directory_cache('switchboard') . '/switchboard.sqlite');
+    if (file_exists(drush_directory_cache('switchboard') . '/switchboard.sqlite')) {
+      unlink(drush_directory_cache('switchboard') . '/switchboard.sqlite');
+    }
   }
   public static function siteGet() {
     static $pdo;
